backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic[email]==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
authlib==1.3.2
httpx==0.27.2
